--- a/tests/runner/Cargo.toml
+++ b/tests/runner/Cargo.toml
@@ -18,6 +18,8 @@ serde = "1.0"
 serde_json = "1.0"
 inventory = "0.1"
 tokio = { version = "1.4", features = ["macros"] }
+bencher = "0.1"
+prettytable-rs = "0.8"
 
 # Fluvio dependencies
 fluvio = { path = "../../src/client" }