--- a/tests/runner/src/main.rs
+++ b/tests/runner/src/main.rs
@@ -1,4 +1,6 @@
 use std::sync::Arc;
+use std::process::exit;
+use std::time::Duration;
 use structopt::StructOpt;
 use fluvio::Fluvio;
 use fluvio_test_util::test_meta::{BaseCli, TestCase, TestCli, TestOption, TestResult};
@@ -8,6 +10,8 @@ use fluvio_future::task::run_block_on;
 use std::panic::{self, AssertUnwindSafe};
 use fluvio_test_util::test_runner::FluvioTest;
 use fluvio_test_util::test_meta::TestTimer;
+use bencher::bench;
+use prettytable::{table, row, cell};
 
 // This is important for `inventory` crate
 #[allow(unused_imports)]
@@ -18,7 +22,7 @@ fn main() {
         let option = BaseCli::from_args();
         //println!("{:?}", option);
 
-        let test_name = option.test_name.clone();
+        let test_name = option.environment.test_name.clone();
 
         // Get test from inventory
         let test =
@@ -45,59 +49,101 @@ fn main() {
         // Deploy a cluster
         let fluvio_client = cluster_setup(&option.environment).await;
 
-        // Create a TestCase object with option.envronment and test_opt
-        let test_case = TestCase::new(option.environment.clone(), test_opt);
+        // Select test
+        let test = inventory::iter::<FluvioTest>
+            .into_iter()
+            .find(|t| t.name == test_name.as_str())
+            .expect("Test not found");
+
+        // Check on test requirements before running the test
+        if !FluvioTest::is_env_acceptable(
+            &(test.requirements)(),
+            &TestCase::new(option.environment.clone(), test_opt.clone()),
+        ) {
+            exit(-1);
+        }
 
         let panic_timer = TestTimer::start();
-
-        let test_result = panic::catch_unwind(AssertUnwindSafe(move || {
-            let test = inventory::iter::<FluvioTest>
-                .into_iter()
-                .find(|t| t.name == test_name.as_str())
-                .expect("Test not found");
-
-            // Run the test
-            (test.test_fn)(fluvio_client.clone(), test_case)
-        }));
-
-        // Handle panics from user tests
         std::panic::set_hook(Box::new(move |_panic_info| {
-            let mut inside_timer = panic_timer.clone();
-            inside_timer.stop();
+            let mut panic_timer = panic_timer.clone();
+            panic_timer.stop();
 
             let test_result = TestResult {
                 success: false,
-                duration: inside_timer.duration(),
+                duration: panic_timer.duration(),
             };
             //run_block_on(async { cluster_cleanup(panic_options.clone()).await });
             eprintln!("Test panicked:\n");
             eprintln!("{}", test_result);
         }));
 
-        // Cluster cleanup
-        cluster_cleanup(option.environment).await;
+        if option.environment.benchmark {
+            run_benchmark(test, fluvio_client, option.environment.clone(), test_opt)
+        } else {
+            let test_result =
+                run_test(option.environment.clone(), test_opt, test, fluvio_client).await;
+            cluster_cleanup(option.environment).await;
 
-        match test_result {
-            // Pass/Fail results
-            Ok(results) => {
-                // We always return TestResults. This is awkward.
-                match results {
-                    Ok(r) => println!("{}", r),
-                    Err(r) => eprintln!("{}", r),
-                }
-            }
-            // Panic results
-            Err(fail_results) => {
-                let test_result = fail_results
-                    .downcast_ref::<TestResult>()
-                    .expect("Failed to convert error from panic")
-                    .to_owned();
-
-                eprintln!("{}", test_result);
-                std::process::exit(-1);
-            }
+            println!("{}", test_result)
         }
     });
+}
+
+async fn run_test(
+    environment: EnvironmentSetup,
+    test_opt: Box<dyn TestOption>,
+    test: &FluvioTest,
+    fluvio_client: Arc<Fluvio>,
+) -> TestResult {
+    let test_case = TestCase::new(environment, test_opt);
+    let test_result = panic::catch_unwind(AssertUnwindSafe(move || {
+        (test.test_fn)(fluvio_client.clone(), test_case)
+    }))
+    .expect("Panic hook should have caught this");
+
+    test_result.expect("Test Result")
+}
+
+fn run_benchmark(
+    test_fn: &FluvioTest,
+    client: Arc<Fluvio>,
+    env: EnvironmentSetup,
+    opts: Box<dyn TestOption>,
+) {
+    println!("Starting benchmark test (Usual test output may be silenced)");
+
+    bench::run_once(move |b| {
+        let summary = b.auto_bench(move |inner_b| {
+            let test_case = TestCase::new(env.clone(), opts.clone());
+            inner_b.iter(|| (test_fn.test_fn)(client.clone(), test_case.clone()))
+        });
+
+        let elapsed_duration = format!("{:?}", Duration::from_nanos(b.ns_elapsed()));
+        let iter_duration = format!("{:?}", Duration::from_nanos(b.ns_per_iter()));
+
+        let table = table!(
+            [b->"Perf Test Summary", "Measurement"],
+            ["Total time elapsed", elapsed_duration],
+            ["Total time elapsed (ns)", b.ns_elapsed()],
+            ["Time per iteration", iter_duration],
+            ["Time per iteration (ns)", b.ns_per_iter()],
+            ["# of iteration", (b.ns_elapsed() / b.ns_per_iter())],
+            ["Sum (ns)", summary.sum],
+            ["Min (ns)", summary.min],
+            ["Max (ns)", summary.max],
+            ["Mean (ns)", summary.mean],
+            ["Median (ns)", summary.median],
+            ["Variance", summary.var],
+            ["Standard Deviation", summary.std_dev],
+            ["Standard Deviation (%)", summary.std_dev_pct],
+            ["Median Absolute Deviation", summary.median_abs_dev],
+            ["Median Absolute Deviation (%)", summary.median_abs_dev_pct],
+            ["Quartiles", format!("{:?}",summary.quartiles)],
+            ["Interquartile Range", summary.iqr]
+        );
+
+        println!("{}", table)
+    })
 }
 
 async fn cluster_cleanup(option: EnvironmentSetup) {
@@ -135,16 +181,11 @@ async fn cluster_setup(option: &EnvironmentSetup) -> Arc<Fluvio> {
 mod tests {
     // CLI Tests
 
-    // Extra vars
-    // No test name
-    // Invalid test name
-    // skip cluster delete then skip cluster start
-    // set topic name
-    // set # spu
-
     use structopt::StructOpt;
     use fluvio_test_util::test_meta::{BaseCli, TestCli};
+    use fluvio_test_util::test_meta::environment::EnvDetail;
     use flv_test::tests::smoke::SmokeTestOption;
+    use std::time::Duration;
 
     #[test]
     fn valid_test_name() {
@@ -172,7 +213,7 @@ mod tests {
 
         if let Some(TestCli::Args(cmd)) = args.test_cmd_args {
             // Structopt commands expect an arbitrary binary name before the test args
-            let mut subcommand = vec![args.test_name.clone()];
+            let mut subcommand = vec![args.environment.test_name.clone()];
             subcommand.extend(cmd);
 
             let smoke_test_case = SmokeTestOption::from_iter(subcommand);
@@ -206,6 +247,20 @@ mod tests {
         let args = BaseCli::from_iter(vec!["flv-test", "smoke", "--spu", "5"]);
 
         assert_eq!(args.environment.spu, 5);
+    }
+
+    #[test]
+    fn timeout() {
+        let args = BaseCli::from_iter(vec!["flv-test", "smoke", "--timeout", "9000"]);
+
+        assert_eq!(args.environment.timeout(), Duration::from_secs(9000));
+    }
+
+    #[test]
+    fn benchmark() {
+        let args = BaseCli::from_iter(vec!["flv-test", "smoke", "--benchmark"]);
+
+        assert!(args.environment.is_benchmark());
     }
 
     //// We validate that the behavior of cluster_setup and cluster_cleanup work as expected