--- a/tests/runner/src/tests/concurrent/mod.rs
+++ b/tests/runner/src/tests/concurrent/mod.rs
@@ -9,6 +9,7 @@ use structopt::StructOpt;
 use fluvio::Fluvio;
 use fluvio_future::task::spawn;
 use fluvio_integration_derive::fluvio_test;
+use fluvio_test_util::test_meta::derive_attr::TestRequirements;
 use fluvio_test_util::test_meta::environment::EnvironmentSetup;
 use fluvio_test_util::test_meta::{TestOption, TestCase, TestResult};