--- a/tests/runner/src/tests/smoke/mod.rs
+++ b/tests/runner/src/tests/smoke/mod.rs
@@ -8,6 +8,7 @@ use structopt::StructOpt;
 
 use fluvio::Fluvio;
 use fluvio_integration_derive::fluvio_test;
+use fluvio_test_util::test_meta::derive_attr::TestRequirements;
 use fluvio_test_util::test_meta::environment::EnvironmentSetup;
 use fluvio_test_util::test_meta::{TestOption, TestCase, TestResult};