--- a/tests/runner/src/utils/Cargo.toml
+++ b/tests/runner/src/utils/Cargo.toml
@@ -19,6 +19,7 @@ proc-macro2 = "1.0"
 inventory = "0.1"
 prettytable-rs = "0.8"
 once_cell = "1.7.2"
+dyn-clone = "1.0"
 
 fluvio = { path = "../../../../src/client" }
 fluvio-future = { version = "0.2.0", features = ["task", "timer", "subscriber", "fixture"] }