--- a/tests/runner/src/utils/test_meta/derive_attr.rs
+++ b/tests/runner/src/utils/test_meta/derive_attr.rs
@@ -2,14 +2,16 @@ use serde::{Serialize, Deserialize};
 use syn::{AttributeArgs, Error as SynError, Lit, Meta, NestedMeta, Path, Result as SynResult};
 use syn::spanned::Spanned;
 use crate::setup::environment::EnvironmentType;
+use std::time::Duration;
 
 #[derive(Debug)]
 pub enum TestRequirementAttribute {
     MinSpu(u16),
     Topic(String),
-    Timeout(u16),
+    Timeout(Duration),
     ClusterType(EnvironmentType),
     TestName(String),
+    Benchmark(bool),
 }
 
 impl TestRequirementAttribute {
@@ -21,6 +23,7 @@ impl TestRequirementAttribute {
                 "timeout" => Self::timeout(&name_value),
                 "cluster_type" => Self::cluster_type(&name_value),
                 "name" => Self::name(&name_value),
+                "benchmark" => Self::benchmark(&name_value),
                 _ => Err(SynError::new(name_value.span(), "Unsupported key")),
             },
             _ => Err(SynError::new(meta.span(), "Unsupported attribute:")),
@@ -35,6 +38,17 @@ impl TestRequirementAttribute {
             .collect();
 
         key.pop().expect("Key expected")
+    }
+
+    fn benchmark(name_value: &syn::MetaNameValue) -> Result<TestRequirementAttribute, SynError> {
+        if let Lit::Bool(b) = &name_value.lit {
+            Ok(Self::Benchmark(b.value()))
+        } else {
+            Err(SynError::new(
+                name_value.span(),
+                "Benchmark must be a LitBool",
+            ))
+        }
     }
 
     fn name(name_value: &syn::MetaNameValue) -> Result<TestRequirementAttribute, SynError> {
@@ -70,9 +84,8 @@ impl TestRequirementAttribute {
 
     fn timeout(name_value: &syn::MetaNameValue) -> Result<TestRequirementAttribute, SynError> {
         if let Lit::Int(timeout) = &name_value.lit {
-            Ok(Self::Timeout(
-                u16::from_str_radix(timeout.base10_digits(), 10).expect("Parse failed"),
-            ))
+            let parsed = u64::from_str_radix(timeout.base10_digits(), 10).expect("Parse failed");
+            Ok(Self::Timeout(Duration::from_secs(parsed)))
         } else {
             Err(SynError::new(name_value.span(), "Timeout must be LitInt"))
         }
@@ -102,9 +115,10 @@ impl TestRequirementAttribute {
 pub struct TestRequirements {
     pub min_spu: Option<u16>,
     pub topic: Option<String>,
-    pub timeout: Option<u16>,
+    pub timeout: Option<Duration>,
     pub cluster_type: Option<EnvironmentType>,
     pub test_name: Option<String>,
+    pub benchmark: Option<bool>,
 }
 
 impl TestRequirements {
@@ -140,6 +154,9 @@ impl TestRequirements {
                 }
                 TestRequirementAttribute::TestName(name) => {
                     test_requirements.test_name = Some(name)
+                }
+                TestRequirementAttribute::Benchmark(is_benchmark) => {
+                    test_requirements.benchmark = Some(is_benchmark)
                 }
             }
         }