Built off of PR #895

* Add benchmarking support utilizing the [bencher](https://crates.io/crates/bencher) crate.
* Add new benchmark supported test `producer_stress`, a single-core producer
* Update README

Resolves #879
Resolves #883 
Resolves #951 
Closes #566 


## Example:
### Without benchmarking
```
$ cargo run --bin flv-test -- producer_stress -d -k -- --producers 3 --iteration 100
    Finished dev [unoptimized + debuginfo] target(s) in 0.16s
     Running `target/debug/flv-test producer_stress -d -k -- --producers 3 --iteration 100`
Start running fluvio test runner
skipping cluster start
Creating the topic: test
topic "test" already exists

Starting single-process producer stress
Producers              : 3
# messages per producer: 100
skipping cluster delete

+--------------+--------------+
| Test Results |              |
+--------------+--------------+
| Pass?        | true         |
+--------------+--------------+
| Duration     | 289.408879ms |
+--------------+--------------+
```

### With benchmarking

```
$ cargo run --bin flv-test -- producer_stress -d -k --benchmark -- --producers 3 --iteration 100
    Finished dev [unoptimized + debuginfo] target(s) in 0.17s
     Running `target/debug/flv-test producer_stress -d -k --benchmark -- --producers 3 --iteration 100`
Start running fluvio test runner
skipping cluster start
Starting benchmark test (Usual test output may be silenced)
+-------------------------------+------------------------------------------+
| Perf Test Summary             | Measurement                              |
+-------------------------------+------------------------------------------+
| Total time elapsed            | 1.402356271s                             |
+-------------------------------+------------------------------------------+
| Total time elapsed (ns)       | 1402356271                               |
+-------------------------------+------------------------------------------+
| Time per iteration            | 280.471254ms                             |
+-------------------------------+------------------------------------------+
| Time per iteration (ns)       | 280471254                                |
+-------------------------------+------------------------------------------+
| # of iteration                | 5                                        |
+-------------------------------+------------------------------------------+
| Sum (ns)                      | 14306013620.45                           |
+-------------------------------+------------------------------------------+
| Min (ns)                      | 281061233.6                              |
+-------------------------------+------------------------------------------+
| Max (ns)                      | 290169616.55                             |
+-------------------------------+------------------------------------------+
| Mean (ns)                     | 286120272.40900004                       |
+-------------------------------+------------------------------------------+
| Median (ns)                   | 286370877.5                              |
+-------------------------------+------------------------------------------+
| Variance                      | 7845839165779.815                        |
+-------------------------------+------------------------------------------+
| Standard Deviation            | 2801042.5140971737                       |
+-------------------------------+------------------------------------------+
| Standard Deviation (%)        | 0.978973803748226                        |
+-------------------------------+------------------------------------------+
| Median Absolute Deviation     | 2930262.531                              |
+-------------------------------+------------------------------------------+
| Median Absolute Deviation (%) | 1.023240406490007                        |
+-------------------------------+------------------------------------------+
| Quartiles                     | (284391694.25, 286370877.5, 288307360.5) |
+-------------------------------+------------------------------------------+
| Interquartile Range           | 3915666.25                               |
+-------------------------------+------------------------------------------+
```