@tjtelan 2021-04-15
> This PR seems to combine multiple features (upgrade, hourly, ...) even thought they are unrelated?



> This PR seems to combine multiple features (upgrade, hourly, ...) even thought they are unrelated?

I removed hourly and upgrade into their own PRs, but they depend on this one to be merged.

https://github.com/infinyon/fluvio/pull/963
https://github.com/infinyon/fluvio/pull/964