@tjtelan 2021-04-16
> Is this PR still relevant now?

Yes, I need this to merge before #963