@sehz 2021-04-16
Can't test this locally due to: https://github.com/infinyon/fluvio/issues/969